Refine macOS dashboard overlay performance
