+++ packages/macos-dashboard/Sources/ClassroomWidgetsDashboard/KeyboardShortcutRecorder.swift
@@ -125,6 +125,10 @@ private final class RecorderNSView: NSView {
 
     override func keyDown(with event: NSEvent) {
         guard isRecording else {
+            if event.keyCode == UInt16(kVK_Space) || event.keyCode == UInt16(kVK_Return) {
+                startRecording()
+                return
+            }
             super.keyDown(with: event)
             return
         }
@@ -153,7 +157,7 @@ private final class RecorderNSView: NSView {
         }
 
         let flags = event.modifierFlags.intersection(.deviceIndependentFlagsMask)
-        textField.stringValue = modifierSymbols(from: flags) + "..."
+        textField.stringValue = DashboardShortcutFormatter.modifierSymbols(from: flags) + "..."
         textField.textColor = .labelColor
     }
 
@@ -175,6 +179,9 @@ private final class RecorderNSView: NSView {
     }
 
     private func setup() {
+        setAccessibilityRole(.button)
+        setAccessibilityLabel("Keyboard shortcut")
+        setAccessibilityHelp("Press to record a new keyboard shortcut")
         addSubview(textField)
         textField.translatesAutoresizingMaskIntoConstraints = false
         NSLayoutConstraint.activate([
@@ -191,6 +198,7 @@ private final class RecorderNSView: NSView {
         window?.makeFirstResponder(self)
         textField.stringValue = "Press shortcut..."
         textField.textColor = .labelColor
+        setAccessibilityValue("Recording shortcut")
         delegate?.recorderDidStartRecording()
     }
 
@@ -205,53 +213,13 @@ private final class RecorderNSView: NSView {
         if currentKeyCode == -1 {
             textField.stringValue = placeholder
             textField.textColor = .secondaryLabelColor
+            setAccessibilityValue("No shortcut")
             return
         }
 
-        let flags = NSEvent.ModifierFlags(rawValue: UInt(currentModifiers))
-        textField.stringValue = modifierSymbols(from: flags) + keyString(from: currentKeyCode)
+        textField.stringValue = DashboardShortcutFormatter.title(keyCode: currentKeyCode, modifiers: currentModifiers) ?? "?"
         textField.textColor = .labelColor
-    }
-
-    private func modifierSymbols(from flags: NSEvent.ModifierFlags) -> String {
-        var symbols = ""
-        if flags.contains(.control) { symbols += "⌃" }
-        if flags.contains(.option) { symbols += "⌥" }
-        if flags.contains(.shift) { symbols += "⇧" }
-        if flags.contains(.command) { symbols += "⌘" }
-        return symbols
-    }
-
-    private func keyString(from keyCode: Int) -> String {
-        let specialKeys: [Int: String] = [
-            kVK_Return: "↩",
-            kVK_Tab: "⇥",
-            kVK_Space: "Space",
-            kVK_Delete: "⌫",
-            kVK_ForwardDelete: "⌦",
-            kVK_Escape: "⎋",
-            kVK_LeftArrow: "←",
-            kVK_RightArrow: "→",
-            kVK_UpArrow: "↑",
-            kVK_DownArrow: "↓",
-            kVK_Home: "↖",
-            kVK_End: "↘",
-            kVK_PageUp: "⇞",
-            kVK_PageDown: "⇟",
-            kVK_F1: "F1", kVK_F2: "F2", kVK_F3: "F3", kVK_F4: "F4",
-            kVK_F5: "F5", kVK_F6: "F6", kVK_F7: "F7", kVK_F8: "F8",
-            kVK_F9: "F9", kVK_F10: "F10", kVK_F11: "F11", kVK_F12: "F12"
-        ]
-
-        if let specialKey = specialKeys[keyCode] {
-            return specialKey
-        }
-
-        guard let scalar = keyCodeToUSKeyboardCharacter[keyCode] else {
-            return "?"
-        }
-
-        return scalar
+        setAccessibilityValue(textField.stringValue)
     }
 
     private static let modifierOnlyKeys: Set<UInt16> = [
@@ -261,19 +229,4 @@ private final class RecorderNSView: NSView {
         UInt16(kVK_Shift), UInt16(kVK_RightShift)
     ]
 
-    private let keyCodeToUSKeyboardCharacter: [Int: String] = [
-        kVK_ANSI_A: "A", kVK_ANSI_S: "S", kVK_ANSI_D: "D", kVK_ANSI_F: "F",
-        kVK_ANSI_H: "H", kVK_ANSI_G: "G", kVK_ANSI_Z: "Z", kVK_ANSI_X: "X",
-        kVK_ANSI_C: "C", kVK_ANSI_V: "V", kVK_ANSI_B: "B", kVK_ANSI_Q: "Q",
-        kVK_ANSI_W: "W", kVK_ANSI_E: "E", kVK_ANSI_R: "R", kVK_ANSI_Y: "Y",
-        kVK_ANSI_T: "T", kVK_ANSI_1: "1", kVK_ANSI_2: "2", kVK_ANSI_3: "3",
-        kVK_ANSI_4: "4", kVK_ANSI_6: "6", kVK_ANSI_5: "5", kVK_ANSI_Equal: "=",
-        kVK_ANSI_9: "9", kVK_ANSI_7: "7", kVK_ANSI_Minus: "-", kVK_ANSI_8: "8",
-        kVK_ANSI_0: "0", kVK_ANSI_RightBracket: "]", kVK_ANSI_O: "O",
-        kVK_ANSI_U: "U", kVK_ANSI_LeftBracket: "[", kVK_ANSI_I: "I",
-        kVK_ANSI_P: "P", kVK_ANSI_L: "L", kVK_ANSI_J: "J", kVK_ANSI_Quote: "'",
-        kVK_ANSI_K: "K", kVK_ANSI_Semicolon: ";", kVK_ANSI_Backslash: "\\",
-        kVK_ANSI_Comma: ",", kVK_ANSI_Slash: "/", kVK_ANSI_N: "N",
-        kVK_ANSI_M: "M", kVK_ANSI_Period: "."
-    ]
 }
