+++ packages/teacher/src/features/desktop/LiquidGlassOverlay.tsx
@@ -17,39 +17,61 @@ const LiquidGlassOverlay: React.FC<LiquidGlassOverlayProps> = ({ active }) => {
     const context = canvas.getContext('2d');
     if (!context) return;
 
-    let frame = 0;
+    const frameInterval = 1000 / 30;
+    const reducedMotion = window.matchMedia('(prefers-reduced-motion: reduce)').matches;
+    let frame: number | null = null;
+    let lastFrameTime = 0;
+    let width = 0;
+    let height = 0;
     let pointerX = 0.5;
     let pointerY = 0.5;
     let hover = 0;
+    let targetHover = 0;
 
     const resize = () => {
       const rect = parent.getBoundingClientRect();
       const ratio = window.devicePixelRatio || 1;
+      width = rect.width;
+      height = rect.height;
       canvas.width = Math.max(1, Math.round(rect.width * ratio));
       canvas.height = Math.max(1, Math.round(rect.height * ratio));
       canvas.style.width = `${rect.width}px`;
       canvas.style.height = `${rect.height}px`;
       context.setTransform(ratio, 0, 0, ratio, 0, 0);
+      draw(performance.now());
     };
 
     const handlePointerMove = (event: PointerEvent) => {
       const rect = parent.getBoundingClientRect();
       pointerX = (event.clientX - rect.left) / Math.max(1, rect.width);
       pointerY = (event.clientY - rect.top) / Math.max(1, rect.height);
-      hover = 1;
+      targetHover = 1;
+      if (reducedMotion) {
+        hover = 1;
+        draw(performance.now());
+      } else {
+        scheduleFrame();
+      }
     };
 
     const handlePointerLeave = () => {
-      hover = 0;
+      targetHover = 0;
+      if (reducedMotion) {
+        hover = 0;
+        draw(performance.now());
+      } else {
+        scheduleFrame();
+      }
     };
 
     const draw = (time: number) => {
-      const rect = parent.getBoundingClientRect();
-      const width = rect.width;
-      const height = rect.height;
+      if (width <= 0 || height <= 0) {
+        return;
+      }
+
       context.clearRect(0, 0, width, height);
 
-      const radius = Math.min(width, height) * (0.25 + hover * 0.18);
+      const radius = Math.max(1, Math.min(width, height) * (0.25 + hover * 0.18));
       const centerX = pointerX * width;
       const centerY = pointerY * height;
       const pulse = Math.sin(time / 550) * 0.5 + 0.5;
@@ -90,22 +112,49 @@ const LiquidGlassOverlay: React.FC<LiquidGlassOverlayProps> = ({ active }) => {
         context.stroke();
       }
       context.globalAlpha = 1;
-
-      frame = window.requestAnimationFrame(draw);
     };
+
+    const animate = (time: number) => {
+      frame = null;
+
+      if (time - lastFrameTime < frameInterval) {
+        scheduleFrame();
+        return;
+      }
+
+      lastFrameTime = time;
+      hover += (targetHover - hover) * 0.22;
+
+      if (Math.abs(targetHover - hover) < 0.01) {
+        hover = targetHover;
+      }
+
+      draw(time);
+
+      if (targetHover > 0 || hover > 0) {
+        scheduleFrame();
+      }
+    };
+
+    function scheduleFrame() {
+      if (!reducedMotion && frame === null) {
+        frame = window.requestAnimationFrame(animate);
+      }
+    }
 
     resize();
     const observer = new ResizeObserver(resize);
     observer.observe(parent);
     parent.addEventListener('pointermove', handlePointerMove);
     parent.addEventListener('pointerleave', handlePointerLeave);
-    frame = window.requestAnimationFrame(draw);
 
     return () => {
       observer.disconnect();
       parent.removeEventListener('pointermove', handlePointerMove);
       parent.removeEventListener('pointerleave', handlePointerLeave);
-      window.cancelAnimationFrame(frame);
+      if (frame !== null) {
+        window.cancelAnimationFrame(frame);
+      }
     };
   }, [active]);
